requirements.txt
beir==2.0.0
cleanlab==2.5.0
cleanvision==0.3.6
colorama==0.4.6
datasets==2.16.0
demjson3==3.0.6
django_haystack==3.2.1
HEBO==0.3.5
Jinja2==3.1.2
jsonschema==4.21.1
loguru==0.7.2
matplotlib==3.8.3
networkx==3.2.1
numpy==1.24.4
pandas==2.2.2
pygraphviz==1.12
python-dotenv==1.0.1
PyYAML==6.0.1
PyYAML==6.0.1
regex==2023.10.3
scikit_learn==1.3.2
scipy==1.13.0
setuptools==67.6.1
tqdm==4.66.1
umap_learn==0.5.5
xgboost==2.0.3
